refactor(common): sample app for component reuse
