+++ package.json
@@ -16,6 +16,7 @@
     "classnames": "^2.2.6",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "react-table": "^7.2.2",
     "typeface-roboto": "^0.0.75",
@@ -75,6 +76,7 @@
     "@types/lingui__core": "^2.7.0",
     "@types/lingui__macro": "^2.7.3",
     "@types/lingui__react": "^2.8.2",
+    "@types/react-router-dom": "^5.1.5",
     "@types/react-table": "^7.0.19",
     "babel-core": "^7.0.0-bridge.0",
     "husky": "^4.2.5",
